+++ src/commands/generate/index.ts
@@ -1,0 +1,2 @@
+export * from './action'
+export * from './routes'
