+++ src/commands/pull/action.ts
@@ -1,6 +1,6 @@
-import { foreachApp } from '../lib/apps/foreach'
-import { readCLIConfig } from '../lib/config'
-import { hideStatus, showStatus } from '../model/status'
+import { foreachApp } from '../../lib/apps/foreach'
+import { readCLIConfig } from '../../lib/config'
+import { hideStatus, showStatus } from '../../model/status'
 
 export async function pull() {
   const { apps } = await readCLIConfig()
