+++ src/commands/start/action.ts
