+++ src/commands/start/model/logs.ts
