+++ src/model/routing.ts
@@ -1,18 +1,11 @@
-import { Command } from 'commander'
 import { createEvent, restore } from 'effector'
+import { Route } from '../types'
 import { executionSucceed } from './lifecycle'
-
-export type RouteConfig = {
-  name: string
-  command: Command
-  component: () => JSX.Element
-  action: () => void | Promise<void>
-}
 
 const routeChanged = createEvent<string>()
 export const $activeRoute = restore(routeChanged, null)
 
-export function prepareRoutes(routes: RouteConfig[]) {
+export function prepareRoutes(routes: Route[]) {
   for (const route of routes) {
     route.command.action(async () => {
       routeChanged(route.name)
