+++ src/ui/screens/start.tsx
@@ -2,7 +2,7 @@ import { useStore } from 'effector-react'
 import { Box, Text } from 'ink'
 import TextInput from 'ink-text-input'
 import React, { useState } from 'react'
-import { $appSummary, commandReceived } from '../../actions/start/model'
+import { $appSummary, commandReceived } from '../../commands/start/model'
 import { BoxNewline } from '../lib'
 import { Table } from '../lib/components/table'
 import { DefaultTemplate } from '../templates'
